refactor loader and error display
